src/inference/requirements.txt
mlflow
pyspark
python-dotenv
xgboost
grpcio>=1.48.1
googleapis-common-protos>=1.56.2
imblearn
scikit-learn
dill
numpy

mlflow
boto3
